+++ src/bin/investigate.ts
@@ -1,7 +1,7 @@
 import 'source-map-support/register';
 import { engineConfigFromENV } from '../lib/config/env';
-import { DataSet, dataSetConfigFromENV } from '../lib/data/data';
 import { dataManager } from '../lib/data/manager';
+import { DataSet, dataSetConfigFromENV } from '../lib/data/set';
 import { Engine } from "../lib/engine";
 import { printSummary } from "../lib/engine/summary";
