+++ src/lib/engine/index.ts
@@ -2,8 +2,8 @@ import chalk from "chalk";
 import { ContactGenerator } from "../contact-generator";
 import { identifyAndFlagContactTypes } from "../contact-generator/contact-types";
 import { updateContactsBasedOnMatchResults } from "../contact-generator/update-contacts";
-import { DataSet } from "../data/data";
 import { RawDataSet } from "../data/raw";
+import { DataSet } from "../data/set";
 import { DealGenerator } from "../deal-generator";
 import { LicenseGrouper } from "../license-matching/license-grouper";
 import { LogDir } from "../log";
